Move vcr_test_base to testsdk
