+++ src/azure-cli-core/tests/test_vcr_security.py
@@ -4,13 +4,12 @@
 # --------------------------------------------------------------------------------------------
 
 # pylint: disable=line-too-long
-import os
 import unittest
 
 
 class Test_vcr_security(unittest.TestCase):
     def test_deployment_name_scrub(self):
-        from azure.cli.core.test_utils.vcr_test_base import _scrub_deployment_name as scrub_deployment_name
+        from azure.cli.testsdk.vcr_test_base import _scrub_deployment_name as scrub_deployment_name
         uri1 = 'https://www.contoso.com/deployments/azurecli1466174372.33571889479?api-version=2015-11-01'
         uri2 = 'https://www.contoso.com/deployments/azurecli1466174372.33571889479/more'
