+++ src/command_modules/azure-cli-role/tests/test_role.py
@@ -9,7 +9,7 @@ import os
 import tempfile
 import time
 
-from azure.cli.core.test_utils.vcr_test_base import VCRTestBase, JMESPathCheck, \
+from azure.cli.testsdk.vcr_test_base import VCRTestBase, JMESPathCheck, \
     ResourceGroupVCRTestBase, NoneCheck, MOCKED_SUBSCRIPTION_ID
